Implemented CreateTokenAttachSourceEffect and refactored existing cards to use it
