+++ Mage.Sets/src/mage/cards/d/DraugrsHelm.java
@@ -6,7 +6,7 @@ import mage.abilities.Ability;
 import mage.abilities.common.EntersBattlefieldTriggeredAbility;
 import mage.abilities.common.SimpleStaticAbility;
 import mage.abilities.costs.mana.ManaCostsImpl;
-import mage.abilities.effects.common.CreateTokenEffect;
+import mage.abilities.effects.common.CreateTokenAttachSourceEffect;
 import mage.abilities.effects.common.DoIfCostPaid;
 import mage.abilities.effects.common.continuous.BoostEquippedEffect;
 import mage.abilities.effects.common.continuous.GainAbilityAttachedEffect;
@@ -17,8 +17,6 @@ import mage.constants.SubType;
 import mage.cards.CardImpl;
 import mage.cards.CardSetInfo;
 import mage.constants.CardType;
-import mage.game.Game;
-import mage.game.permanent.Permanent;
 import mage.game.permanent.token.ZombieBerserkerToken;
 
 /**
@@ -35,7 +33,7 @@ public final class DraugrsHelm extends CardImpl {
         // When Draugr's Helm enters the battlefield, you may pay {2}{B}.
         // If you do, create a 2/2 black Zombie Berserker creature token, then attach Draugr's Helm to it.
         this.addAbility(new EntersBattlefieldTriggeredAbility(new DoIfCostPaid(
-                new DraugrsHelmEffect(), new ManaCostsImpl<>("{2}{B}")
+                new CreateTokenAttachSourceEffect(new ZombieBerserkerToken()), new ManaCostsImpl<>("{2}{B}")
         )));
 
         // Equipped creature gets +2/+2 and has menace.
@@ -54,33 +52,5 @@ public final class DraugrsHelm extends CardImpl {
     @Override
     public DraugrsHelm copy() {
         return new DraugrsHelm(this);
-    }
-}
-
-class DraugrsHelmEffect extends CreateTokenEffect {
-
-    public DraugrsHelmEffect() {
-        super(new ZombieBerserkerToken());
-        staticText = "create a 2/2 black Zombie Berserker creature token, then attach {this} to it";
-    }
-
-    private DraugrsHelmEffect(final DraugrsHelmEffect effect) {
-        super (effect);
-    }
-
-    @Override
-    public boolean apply (Game game, Ability source) {
-        super.apply(game, source);
-        Permanent token = game.getPermanent(this.getLastAddedTokenId());
-        if (token != null) {
-            token.addAttachment(source.getSourceId(), source, game);
-            return true;
-        }
-        return false;
-    }
-
-    @Override
-    public DraugrsHelmEffect copy() {
-        return new DraugrsHelmEffect(this);
     }
 }
