+++ Mage.Sets/src/mage/cards/w/WolfridersSaddle.java
@@ -3,7 +3,7 @@ package mage.cards.w;
 import mage.abilities.Ability;
 import mage.abilities.common.EntersBattlefieldTriggeredAbility;
 import mage.abilities.common.SimpleStaticAbility;
-import mage.abilities.effects.common.CreateTokenEffect;
+import mage.abilities.effects.common.CreateTokenAttachSourceEffect;
 import mage.abilities.effects.common.combat.CantBeBlockedByMoreThanOneSourceEffect;
 import mage.abilities.effects.common.continuous.BoostEquippedEffect;
 import mage.abilities.effects.common.continuous.GainAbilityAttachedEffect;
@@ -13,10 +13,7 @@ import mage.cards.CardSetInfo;
 import mage.constants.AttachmentType;
 import mage.constants.CardType;
 import mage.constants.SubType;
-import mage.game.Game;
-import mage.game.permanent.Permanent;
 import mage.game.permanent.token.WolfToken;
-import mage.players.Player;
 
 import java.util.UUID;
 
@@ -31,7 +28,7 @@ public final class WolfridersSaddle extends CardImpl {
         this.subtype.add(SubType.EQUIPMENT);
 
         // When Wolfrider's Saddle enters the battlefield, create a 2/2 green Wolf creature token, then attach Wolfrider's Saddle to it.
-        this.addAbility(new EntersBattlefieldTriggeredAbility(new WolfridersSaddleEffect()));
+        this.addAbility(new EntersBattlefieldTriggeredAbility(new CreateTokenAttachSourceEffect(new WolfToken())));
 
         // Equipped creature gets +1/+1 and can't be blocked by more than one creature.
         Ability ability = new SimpleStaticAbility(new BoostEquippedEffect(1, 1));
@@ -51,36 +48,5 @@ public final class WolfridersSaddle extends CardImpl {
     @Override
     public WolfridersSaddle copy() {
         return new WolfridersSaddle(this);
-    }
-}
-
-class WolfridersSaddleEffect extends CreateTokenEffect {
-
-    WolfridersSaddleEffect() {
-        super(new WolfToken());
-        staticText = "create a 2/2 green Wolf creature token, then attach {this} to it.";
-    }
-
-    private WolfridersSaddleEffect(final WolfridersSaddleEffect effect) {
-        super(effect);
-    }
-
-    @Override
-    public boolean apply(Game game, Ability source) {
-        Player controller = game.getPlayer(source.getControllerId());
-        if (controller == null || !super.apply(game, source)) {
-            return false;
-        }
-        Permanent p = game.getPermanent(this.getLastAddedTokenId());
-        if (p == null) {
-            return false;
-        }
-        p.addAttachment(source.getSourceId(), source, game);
-        return true;
-    }
-
-    @Override
-    public WolfridersSaddleEffect copy() {
-        return new WolfridersSaddleEffect(this);
     }
 }
